refactor: 핸들러 함수 hook으로 분리
