+++ frontend/techpick/src/features/folderManagement/hooks/useGetStructure.ts
@@ -1,0 +1,9 @@
+import { useQuery } from '@tanstack/react-query';
+import { getStructure } from '@/features/queryFunctions';
+
+export const useGetStructure = () => {
+  return useQuery({
+    queryKey: ['getStructure'],
+    queryFn: getStructure,
+  });
+};
