+++ frontend/techpick/src/features/folderManagement/hooks/useMoveFolder.ts
@@ -1,0 +1,8 @@
+import { putFolderMove } from '@/features/queryFunctions';
+import { useMutation } from '@tanstack/react-query';
+
+export const useMoveFolder = () => {
+  return useMutation({
+    mutationFn: putFolderMove,
+  });
+};
