+++ frontend/techpick/src/widgets/DirectoryTreeSection/DirectoryTreeSection.tsx
@@ -14,43 +14,21 @@ import {
 import Image from 'next/image';
 import { ToggleThemeButton } from '@/features/';
 import { NodeData } from '@/shared/types/NodeData';
-import {
-  CreateHandler,
-  MoveHandler,
-  NodeApi,
-  Tree,
-  TreeApi,
-} from 'react-arborist';
+import { NodeApi, Tree, TreeApi } from 'react-arborist';
 import useResizeObserver from 'use-resize-observer';
 import { DirectoryNode } from '@/components';
 import { useDragDropManager } from 'react-dnd';
-import { moveNode } from '@/features/moveNode';
 import { useTreeStore } from '@/shared/stores/treeStore';
 import { Folder } from 'lucide-react';
 import { EditorContextMenu } from '../EditorContextMenu';
-import { useMutation, useQuery } from '@tanstack/react-query';
-import {
-  getFoldersIdData,
-  getStructure,
-  postFolder,
-  putFolderMove,
-} from '@/features/queryFunctions';
-import { createNode } from '@/features/createNode';
-import { DefaultFolderIdData } from '@/shared/types/ApiTypes';
+import { useGetStructure } from '@/features/folderManagement/hooks/useGetStructure';
+import { useTreeHandlers } from '@/features/folderManagement/hooks/useTreeHandlers';
 
 export function DirectoryTreeSection() {
   const { ref, width, height } = useResizeObserver<HTMLDivElement>();
   const treeRef = useRef<TreeApi<NodeData> | undefined>(undefined);
   const dragDropManager = useDragDropManager();
-  const {
-    focusedNode,
-    focusedFolderNodeList,
-    focusedLinkNodeList,
-    setTreeRef,
-    setFocusedFolderNodeList,
-    setFocusedLinkNodeList,
-    setFocusedNode,
-  } = useTreeStore();
+  const { setTreeRef, setFocusedNode } = useTreeStore();
 
   const handleTreeRef = (instance: TreeApi<NodeData> | null | undefined) => {
     if (instance && !treeRef.current) {
@@ -64,126 +42,9 @@ export function DirectoryTreeSection() {
     data: structureData,
     error: structureError,
     isLoading: isStructureLoading,
-    refetch: refetchStructure,
-  } = useQuery({
-    queryKey: ['getStructure'],
-    queryFn: getStructure,
-  });
+  } = useGetStructure();
 
-  // ============== 기본 폴더 id 가져오기 ============
-  const { data: defaultFolderIdData } = useQuery<DefaultFolderIdData>({
-    queryKey: ['getDefaultFolderId'],
-    queryFn: getFoldersIdData,
-  });
-
-  // ==============  새로운 폴더 생성  =============
-  const { mutateAsync: createFolder } = useMutation<
-    {
-      id: number;
-      name: string;
-      parentFolderId: number;
-      userId: number;
-    },
-    Error,
-    string
-  >({
-    mutationFn: postFolder,
-  });
-
-  // ============  폴더 이동 ================
-  const { mutateAsync: moveFolder } = useMutation<
-    void,
-    Error,
-    { folderId: string; structure: object }
-  >({
-    mutationFn: putFolderMove,
-  });
-
-  // ============== Tree 핸들러 함수 ===============
-  const handleCreate: CreateHandler<NodeData> = async ({
-    parentId,
-    parentNode,
-    index,
-    type,
-  }): Promise<{ id: string }> => {
-    try {
-      // 폴더 생성 (서버)
-      const data = await createFolder('New Folder27');
-      console.log('Server: Folder created:', data);
-
-      // 폴더 생성 (클라이언트)
-      const updatedTreeData = createNode(
-        structureData!.root,
-        focusedNode,
-        type,
-        treeRef.current,
-        parentId,
-        parentNode,
-        index,
-        data.id,
-        data.name
-      );
-
-      // 서버에 업데이트된 트리 전송
-      const serverData = {
-        parentFolderId: defaultFolderIdData!.ROOT,
-        structure: {
-          root: updatedTreeData,
-          recycleBin: [],
-        },
-      };
-      console.log('defaultFolderIdData', defaultFolderIdData);
-      console.log('서버용 data', serverData);
-      // 폴더 이동 (서버)
-      await moveFolder({ folderId: data.id.toString(), structure: serverData });
-
-      // 구조 데이터 새로 가져오기
-      refetchStructure();
-    } catch (error) {
-      console.error('Error creating folder:', error);
-    }
-    return { id: '999' };
-  };
-
-  const handleMove: MoveHandler<NodeData> = async ({
-    dragIds,
-    dragNodes,
-    parentId,
-    parentNode,
-    index,
-  }) => {
-    const updatedTreeData = moveNode(
-      structureData!.root,
-      focusedNode,
-      focusedFolderNodeList,
-      focusedLinkNodeList,
-      setFocusedFolderNodeList,
-      setFocusedLinkNodeList,
-      dragIds[0],
-      dragNodes[0],
-      parentId,
-      parentNode,
-      index
-    );
-    // 서버에 업데이트된 트리 전송
-    const serverData = {
-      parentFolderId: parentNode
-        ? parentNode.data.folderId
-        : defaultFolderIdData!.ROOT,
-      structure: {
-        root: updatedTreeData,
-        recycleBin: [],
-      },
-    };
-    console.log('defaultFolderIdData', defaultFolderIdData);
-    console.log('서버에 보낼 데이터', serverData);
-
-    await moveFolder({
-      folderId: dragNodes[0].data.folderId!.toString(),
-      structure: serverData,
-    });
-    refetchStructure();
-  };
+  const { handleCreate, handleMove } = useTreeHandlers();
 
   return (
     <div className={leftSidebarSection}>
